DRY up bench scripts, remove sudo logic, add reliability tips
- Extract shared formatTime, deltaEmoji, parsePairs into bench-utils.mjs
- Simplify run-bench.sh to just CPU pinning (no sudo/nice complexity)
- Simplify bench-compare.mjs to just CPU pinning
- Replace interactive sudo prompt in local-bench-summary.sh with a
  "tips for more reliable results" section printed after results
